refactor: change controllers folder structure
