+++ src/controllers/http-api/routes/FeatsAPIRoutes.ts
@@ -1,8 +1,8 @@
 import { Request } from "express";
-import { Feats } from "../../features/Feats";
-import { IFeats } from "../../features/interfaces/interfaces";
-import { isObjectId } from "../../infra/database/mongodb";
-import { BadRequest } from "../../util/errors/HttpStatusCode";
+import { Feats } from "../../../features/Feats";
+import { IFeats } from "../../../features/interfaces/interfaces";
+import { isObjectId } from "../../../infra/database/mongodb";
+import { BadRequest } from "../../../util/errors/HttpStatusCode";
 import { checkForMissingProperties } from "./utils";
 
 export default [
