+++ src/controllers/http-api/routes/RecordsAPIRoutes.test.ts
@@ -1,8 +1,8 @@
-import { app } from '../../infra/http-server/index';
+import { app } from '../../../infra/http-server/index';
 import request from 'supertest';
-import { initMongoDB, db, closeDb } from '../../infra/database/mongodb';
-import RecordsRepo from '../../repositories/RecordsRepo';
-import { IRecords } from '../../features/interfaces/interfaces';
+import { initMongoDB, db, closeDb } from '../../../infra/database/mongodb';
+import RecordsRepo from '../../../repositories/RecordsRepo';
+import { IRecords } from '../../../features/interfaces/interfaces';
 
 describe('Records HTTP API Routes', () => {
   const dummyRecord: Partial<IRecords> = {
